+++ src/components/mod.rs
@@ -1,7 +1,9 @@
 pub mod boxed;
 pub mod divider;
 pub mod section;
+pub mod label;
 
 pub use boxed::Boxed;
 pub use divider::Divider;
 pub use section::Section;
+pub use label::Label;